Add login page with header and footer
