+++ frontend/src/App.tsx
@@ -1,11 +1,13 @@
 import Footer from "./components/footer/Footer"
 
-import { Header } from "./header/Header";
+import { Login } from "./components/login/Login";
+import { Header } from "./components/header/Header";
 
 function App() {
   return (
     <>
       <Header />
+      <Login />
       <Footer />
     </>
   )
